Fixed relative path issue in sign-certificates caused by path shuffling

+++ commands/sign-certificate.cmd
@@ -50,7 +50,7 @@ openssl x509 -req -days 365 -sha256 -extensions v3_req            \
   -in "${WARDEN_SSL_DIR}/certs/${CERTIFICATE_NAME}.csr.pem"       \
   -out "${WARDEN_SSL_DIR}/certs/${CERTIFICATE_NAME}.crt.pem" 
 
-if [[ "$(docker-compose -p warden -f docker/docker-compose.yml ps -q traefik)" ]]; then
+if [[ "$(cd "${WARDEN_DIR}" && docker-compose -p warden -f docker/docker-compose.yml ps -q traefik)" ]]; then
   echo "==> Restarting traefik"
   "${WARDEN_DIR}/bin/warden" restart traefik
 fi
